[development/allocators] fixing a compiler warning

+++ source/1_engine/rex_engine/include/public/rex_engine/settings/boot_settings.h
@@ -8,7 +8,7 @@ namespace rex
 {
 	struct BootSettings
 	{
-		s32 single_frame_heap_size = 4_kib;
-		s32 scratch_heap_size = 4_kib;
+		s64 single_frame_heap_size = 4_kib;
+		s64 scratch_heap_size = 4_kib;
 	};
 }
